Fix build on Linux by removing erroneous patching.
The removed hunk is definitely wrong, but I can't figure out what it was
trying to do.  I think the patch hunk was accidentally reintroduced in r1.4
after being correctly removed in r1.3.

The reason it is wrong is because it breaks later tests by introducing code
into confdefs.h.  The following tests always break, because they have
duplicate main() definitions.

Build test on NetBSD works because the java support isn't in the suggested
options anyway.