Fix override-as.mk by fixing the path in AS.
This rarely seems to be a problem, but is exposed by building mplayer on a
32-bit x86 platform with USE_CWRAPPERS=yes.  Possibly with Linux tie-ins, my
NetBSD platforms are all 64 bit.

In the failing scenario mplayer attempts (at least at configure time) to use
"as" directly.  This is wrapped by cwrappers, which attempts to execute $AS
explicitly.

The exec fails because $PREFIX/bin/as doesn't exist, and hasn't done for a
long time.

Correct all the tool paths to include the "g" prefix for all these tools.