Remove bogus ONLY_FOR_PLATFORM.
Note that it currently doesn't build at all (AFAICT) so the mere fact
that it doesn't build on some buggix 0.9 release isn't indicative of
much.

If in the future there turn out to be platforms it really doesn't
build for, use BROKEN_ON_PLATFORM, or ONLY_FOR_PLATFORM if it'll
really never work as opposed to nobody feels like bothering to fix it.