Import py34-URLObject-2.4.0 as www/py-URLObject.
URLObject is a utility class for manipulating URLs. The latest
incarnation of this library builds upon the ideas of its predecessor,
but aims for a clearer API, focusing on proper method names over
operator overrides. It's also being developed from the ground up
in a test-driven manner, and has full Sphinx documentation.