Update tex-etex-pkg{,-doc} to 2.3
Version 2.2 adds a check that etex package has not been loaded after
extended registers are already allocated by the LaTeX format, and gives
a warning in that case.

Previously, if etex.sty was loaded too late LaTeX would already have
run out of registers and given an error.

Since the 2015/01/01 LaTeX release, LaTeX will use extended registers.
If etex.sty is to be loaded it must be loaded before extended registers
are used to avoid the two allocation schemes conflicting.

As explained in the warning message, to avoid the warning, either avoid
loading etex package, or load it earlier in the document.

2.3
Technical re-arrangement of the checks introduced in version 2.2.
Moved the conflicting allocation check before the package options are
declared to avoid errors from the package option system in some
contexts.