Import livestreamer-1.12.2 as multimedia/livestreamer.
Packaged for wip by Jonathan Buschmann, tnn, leot, and coypu.

Livestreamer is a command-line utility that pipes video streams from various
services into a video player, such as VLC. The main purpose of Livestreamer
is to allow the user to avoid buggy and CPU heavy flash plugins but still
be able to enjoy various streamed content.